Inserted data values for PaymentTypes in schema
Added extra logic in OrderAmount to account for coupons. Removed auto-increment for OrderTypes and PaymentTypes primary key value following this. Could cause complications if one wanted to see what the value of the order would have been without the coupon or if the coupon only covered only one part of the order. Might address in the future.