Added extra logic for add-on generation
Won't generate add-ons for soft drinks as that makes no real-world sense. Also won't make tayiaki twist a possibilty for anything but tayiaki type items, and will only limit the amount of it to 1 (the add-on represents putting the tayiaki cone into a cup with extra ice cream, this can only be done once).